# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "druthers.io API local",
    "description": "API for druthers.io \u2014 track and rank the movies, TV, books, and games you actually care about.",
    "contact": {
      "name": "Druthers",
      "url": "https://www.druthers.io/"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://api.druthers.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get All Users",
        "description": "Retrieve all users from the database. Must be an admin to view all users.\n\nArgs:\n    db (Session): The database session.\n\nReturns:\n    List: OutUserDisplay: A list of users.",
        "operationId": "get_all_users_v1_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutResponseUserModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "description": "Create a new user in the database.\n\nArgs:\n    request (InUserBase): Must include email,\n    display_name (max length 16 characters), and password.\n    db (Session): The database session.\n\nReturns:\n    List: OutUserDisplay: The newly created user data.",
        "operationId": "create_user_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InUserBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutResponseUserModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InUserBase": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "email",
          "password"
        ],
        "title": "InUserBase",
        "description": "Schema for user input data."
      },
...
      "OutResponseUserModel": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/OutUserDisplay"
            },
            "type": "array",
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "None"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "OutResponseUserModel",
        "description": "Response format for user data."
      },
...
      "OutUserDisplay": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "user_group": {
            "type": "string",
            "title": "User Group"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "email",
          "user_group",
          "created_at",
          "updated_at"
        ],
        "title": "OutUserDisplay",
        "description": "Schema for displaying user data."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  },
...
}